feat : ai 모델화v2
기존 ai 모듈을 init 이랑 실행모듈로 나누었습니다.
