+++ backend/ai_model/ai_init.py
@@ -1,13 +1,8 @@
 import torch
-import cv2
-from PIL import Image
 import torchvision.transforms as T
 import torchvision
-import numpy as np
 from torch import nn
 import torch.nn.functional as F
-import json
-import io
 
 ## init 코드는 #으로 분리해주세요
 #################################
@@ -126,94 +121,3 @@ shinkaiv2_generator.to('cpu')
 hayao_v2_generator = Generator()
 hayao_v2_generator.load_state_dict(hayao_v2_model_weight)
 hayao_v2_generator.to('cpu')
-#################################
-
-print("hello")
-
-def ai_execute(user_id, origin_id, origin_image, style, background):
-    """
-    style: 만화 선택: 1,2,3 : 1: 신카이_v1, 2:신카이_v2, 3:미야자키_v2
-    background: 배경선택 여부 : 1,2,3 : 1: 인물만, 2: 배경만, 3: 둘다.
-    """
-    result_image_path = './result_images/'+user_id+'_'+origin_id+'.jpg'
-    def get_prediction(img, threshold):
-        transform = T.Compose([T.ToTensor()])
-        img = transform(img)
-
-        pred = model([img])
-        pred_score = list(pred[0]['scores'].detach().numpy())
-
-        pred_t = [pred_score.index(x) for x in pred_score if x>threshold][-1]
-
-        masks = (pred[0]['masks']>0.5).detach().cpu().numpy()
-
-        pred_class = [COCO_INSTANCE_CATEGORY_NAMES[i] for i in list(pred[0]['labels'].numpy())]
-        pred_boxes = [[(i[0], i[1]), (i[2], i[3])] for i in list(pred[0]['boxes'].detach().numpy())]
-
-        masks = masks[:pred_t+1]
-        pred_boxes = pred_boxes[:pred_t+1]
-        pred_class = np.array(pred_class[:pred_t+1])
-        return masks, pred_boxes, pred_class
-    
-    is_converted = False
-    try:
-        masks, pred_boxes, pred_class = get_prediction(origin_image,0.7)
-    except Exception as e:
-        return "파일 없음"
-    
-    segmentation = np.zeros_like(masks[0])
-    for i, value in enumerate(pred_class):
-        if value =='person':
-            segmentation += masks[i]
-    
-    segmentation = segmentation[0][:,:,np.newaxis]
-    person_image = origin_image * segmentation
-    landscape_image = origin_image * (1-segmentation)
-    
-    tf = T.ToTensor()
-    pre_image = (tf(origin_image)*2 - 1).unsqueeze(0)
-
-    if style == 1:
-        output = shinkaiv1_generator(pre_image).detach()
-    elif style == 2:
-        output = shinkaiv2_generator(pre_image).detach()
-    elif style == 3:
-        output = hayao_v2_generator(pre_image).detach()
-    else:
-        img_pil = Image.fromarray(origin_image)
-        buffer_ = io.BytesIO()
-        img_pil.save(buffer_, format='PNG')
-        buffer_.seek(0)
-        img_byte = buffer_.read()
-        return (user_id, origin_id, img_byte, style, background, is_converted)
-    
-    output = ( (((output[0].permute((1,2,0)).numpy()+1)/2)*255).astype('uint8') )
-    
-    # 배경 설정
-    if background == 1:
-        output = output * segmentation
-        result_image = output + landscape_image
-    elif background == 2:
-        output = output * (1-segmentation)
-        result_image = output + person_image
-    elif background == 3: 
-        result_image = output
-    else:
-        img_pil = Image.fromarray(origin_image)
-        buffer_ = io.BytesIO()
-        img_pil.save(buffer_, format='PNG')
-        buffer_.seek(0)
-        img_byte = buffer_.read()
-        return (user_id, origin_id, img_byte, style, background, is_converted)
-    
-    is_converted = True
-    
-    result_image = result_image.astype('uint8')
-    img_pil = Image.fromarray(result_image)
-    buffer_ = io.BytesIO()
-    img_pil.save(buffer_, format='PNG')
-    buffer_.seek(0)
-    img_byte = buffer_.read()
-
-    return (user_id, origin_id, img_byte, style, background, is_converted)
-
